requirements.txt
asgiref==3.8.1
class-registry==2.1.2
dj-database-url==2.3.0
Django==5.2.1
django-colorfield==0.14.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
filters==1.3.2
gunicorn==23.0.0
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
regex==2024.11.6
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.2
whitenoise==6.9.0
